Refactor game loop and network components for improved clarity and performance

+++ Game.ECS/Systems/MovementSystem.cs
@@ -16,8 +16,7 @@ public sealed partial class MovementSystem(World world) : GameSystem(world)
         if (vel.Value is { X: 0, Y: 0 }) return;
         
         // Atualiza a rotação com base na direção da velocidade
-        rotation.Value = (int)(Math.Atan2(vel.Value.Y, vel.Value.X) * (180 / Math.PI));
-
+        rotation.Value = (ushort)(Math.Atan2(vel.Value.Y, vel.Value.X) * (180 / Math.PI));
         // Atualiza a posição com base na velocidade e no tempo decorrido
         pos.Value = new Coordinate(
             pos.Value.X + (int)(vel.Value.X * deltaTime),
